+++ src/test/java/com/hedera/logger/RecordFileParserTest.java
@@ -1,4 +1,4 @@
-package com.hedera.parser;
+package com.hedera.logger;
 
 /*-
  * ‌
